Improved error message
When a string argument with choices gets an illegal value, make the error message clearer:
"USERSTRING is not a valid choice" -> "'USERSTRING' is not a valid choice"

When the string choices are normal English words, sometimes the former form is confusing. For example, if the argument value is 'your choice', the resulting message is "your choice is not a valid choice", which looks vague.